make ship_accept responses link to packages
The ::Response::PackageResult objects in a ::Response::ShipmentAccept
now link back to the ::Package objects they relate to. This is based on
the wide-spread assumption that the result elements are in the same
order as the request elements. Everybody seems to rely on this, even if
I can't find an official confirmation.